Look for <fontset> definitions in xml files in /fonts directory
Modified SettingOptionsSkinFontsFiller in addons/Skin.cpp to get the names of available fontsets from any xml files in the skin's /fonts directory, along with from the Font.xml file. Created new routine GetFontsetsFromFile to look for the names of fontsets in an xml file.
Modified LoadFonts in guilib/GUIFontManager to look for fontset definitions from any xml file in the skin's /fonts directory along with from the Font.xml file.